Add lighthouse layers to the tasking map.
